+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Display custom emojis and greentext when posts are viewed by a guest.
+
 ## [3.22.0] - 2025-04-02
 
 ### Added